// pos-Frontend/src/app/model/Transporter.ts
export class Transporter {
  intTransporterId: number = 0;
  varTransporterName: string = '';
  varContactNo: string = '';
  varAddress: string = '';
  varEmail: string = '';
  intCompanyId: number = 0;
  intCreatedBy: number = 0;
  intUpdatedBy: number = 0;
  dtCreationDate: Date = new Date();
  dtUpdationDate: Date = new Date();
}